Signup refactoring 🔨
