--- src/ignitus-Authentication/ignitus-ProfessorSignUp/Components/Constants.js
@@ -1,3 +1,0 @@
-export const logo = 'https://storage.googleapis.com/ignitus_assets/ig-logo/IgnitusLogos_SVG%20Logo.svg';
-export const studentAuth = 'https://storage.googleapis.com/ignitus_assets/ig-assets/Assets_studentAuth.svg';
-export const professorAuth = 'https://storage.googleapis.com/ignitus_assets/ig-assets/Assets_professorAuth.svg';
