--- src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-AuthenticationComponents/SignUp/constants.js
@@ -1,12 +1,0 @@
-export const logo = 'https://storage.googleapis.com/ignitus_assets/ig-logo/IgnitusLogos_SVG%20Logo.svg';
-export const studentAuth = 'https://storage.googleapis.com/ignitus_assets/ig-assets/Assets_studentAuth.svg';
-export const professorAuth = 'https://storage.googleapis.com/ignitus_assets/ig-assets/Assets_professorAuth.svg';
-export const statePayLoad = {
-  email: '',
-  password: '',
-  confirmPassword: '',
-  equalmessage: false,
-  emptyMessage: false,
-  invalidEmail: false,
-  showPassword: false,
-};
