+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-AuthenticationComponents/SignUp/constants.ts
@@ -1,0 +1,12 @@
+export const logo: string = 'https://storage.googleapis.com/ignitus_assets/ig-logo/IgnitusLogos_SVG%20Logo.svg';
+export const studentAuth: string = 'https://storage.googleapis.com/ignitus_assets/ig-assets/Assets_studentAuth.svg';
+export const professorAuth: string = 'https://storage.googleapis.com/ignitus_assets/ig-assets/Assets_professorAuth.svg';
+export const statePayLoad = {
+  email: '',
+  password: '',
+  confirmPassword: '',
+  equalmessage: false,
+  emptyMessage: false,
+  invalidEmail: false,
+  showPassword: false,
+};
