+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-AuthenticationComponents/SignUp/types.ts
@@ -1,0 +1,1 @@
+export {}
